requirements.txt
django==5.2.7
djangorestframework>=3.15.0
python-dotenv>=1.0.0
dj-database-url>=1.0
psycopg2-binary>=2.9
PyYAML>=6.0
joblib>=1.2
scikit-learn>=1.4
numpy>=1.26
gunicorn>=20.1.0

